+++ src/controllers/EntriesController.php
@@ -335,10 +335,14 @@ class EntriesController extends BaseEntriesController
             (Craft::$app->getIsMultiSite() && Craft::$app->getSites()->currentSite->id != $site->id ? '/'.$site->handle : '');
 
         // Can the user delete the entry?
-        $variables['canDeleteEntry'] = $entry->id !== null && (
+        $variables['canDeleteEntry'] = (
+            get_class($entry) === Entry::class &&
+            $entry->id !== null &&
+            (
                 ($entry->authorId == $currentUser->id && $currentUser->can('deleteEntries'.$variables['permissionSuffix'])) ||
                 ($entry->authorId != $currentUser->id && $currentUser->can('deletePeerEntries'.$variables['permissionSuffix']))
-            );
+            )
+        );
 
         // Full page form variables
         $variables['fullPageForm'] = true;